Rename data directory to npm
