+++ npm/packages.js
@@ -4,6 +4,7 @@ module.exports = {
     'react-dom'
   ],
   devDependencies: [
+    'autoprefixer',
     'ava',
     'babel-core',
     'babel-eslint',
@@ -17,12 +18,15 @@ module.exports = {
     'enzyme',
     'eslint',
     'eslint-plugin-react',
+    'extract-text-webpack-plugin',
     'ignore-styles',
     'jsdom',
     'react-addons-test-utils',
     'react-hot-loader',
     'sinon',
     'style-loader',
+    'precss',
+    'postcss-loader',
     'webpack',
     'webpack-dev-server'
   ]
